[simu] test starting a real simulation, pull simulation dev packages

+++ src/swapi/simu/process.go
@@ -38,10 +38,12 @@ type SimProcess struct {
 }
 
 func (sim *SimProcess) Kill() {
-	sim.tailch <- 1
-	sim.quitAll.Wait()
-	sim.cmd.Process.Kill()
-	sim.cmd.Wait()
+	if sim != nil {
+		sim.tailch <- 1
+		sim.quitAll.Wait()
+		sim.cmd.Process.Kill()
+		sim.cmd.Wait()
+	}
 }
 
 func (sim *SimProcess) GetLogPath() string {
